db is done

+++ Design Stuff/Database/database.sql
@@ -6,6 +6,7 @@ baseDamage REAL,
 baseMoveSpeed REAL,
 optionFX INTEGER,
 optionMusic INTEGER,
-optionSoundLevel INTEGER,
+optionSoundLevel REAL,
 optionTilt INTEGER
 );
+
